+++ airbyte-scheduler/src/main/java/io/airbyte/scheduler/worker_run/WorkerRunFactoryUtils.java
@@ -22,39 +22,16 @@
  * SOFTWARE.
  */
 
-package io.airbyte.scheduler;
+package io.airbyte.scheduler.worker_run;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.verify;
+import io.airbyte.workers.process.AirbyteIntegrationLauncher;
+import io.airbyte.workers.process.IntegrationLauncher;
+import io.airbyte.workers.process.ProcessBuilderFactory;
 
-import io.airbyte.config.JobOutput;
-import io.airbyte.scheduler.worker_run.WorkerRun;
-import io.airbyte.workers.Worker;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
+public class WorkerRunFactoryUtils {
 
-class WorkerRunTest {
-
-  private Path path;
-  private Worker<Integer, JobOutput> worker;
-
-  @SuppressWarnings("unchecked")
-  @BeforeEach
-  void setUp() throws IOException {
-    path = Files.createTempDirectory("test").resolve("sub").resolve("sub");
-    worker = Mockito.mock(Worker.class);
-  }
-
-  @Test
-  void name() throws Exception {
-    new WorkerRun(path, 1, worker).call();
-
-    assertTrue(Files.exists(path));
-    verify(worker).run(1, path);
+  public static IntegrationLauncher createLauncher(long jobId, int attempt, final String image, ProcessBuilderFactory pbf) {
+    return new AirbyteIntegrationLauncher(jobId, attempt, image, pbf);
   }
 
 }
