+++ airbyte-scheduler/src/test/java/io/airbyte/scheduler/SchedulerWorkerRunWithEnvironmentFactoryTest.java
@@ -35,31 +35,22 @@ import com.google.common.collect.Lists;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.config.JobCheckConnectionConfig;
 import io.airbyte.config.JobConfig;
-import io.airbyte.config.JobConfig.ConfigType;
 import io.airbyte.config.JobDiscoverCatalogConfig;
 import io.airbyte.config.JobGetSpecConfig;
-import io.airbyte.config.JobOutput;
 import io.airbyte.config.JobResetConnectionConfig;
 import io.airbyte.config.JobSyncConfig;
-import io.airbyte.config.StandardCheckConnectionInput;
-import io.airbyte.config.StandardDiscoverCatalogInput;
-import io.airbyte.config.StandardSyncInput;
 import io.airbyte.config.State;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.ConfiguredAirbyteStream;
 import io.airbyte.protocol.models.SyncMode;
-import io.airbyte.scheduler.worker_run_factories.CheckConnectionWorkerRunFactory;
-import io.airbyte.scheduler.worker_run_factories.DiscoverWorkerRunFactory;
-import io.airbyte.scheduler.worker_run_factories.GetSpecWorkerRunFactory;
-import io.airbyte.scheduler.worker_run_factories.SyncWorkerRunFactories.ResetConnectionWorkerRunFactory;
-import io.airbyte.scheduler.worker_run_factories.SyncWorkerRunFactories.SyncWorkerRunFactory;
-import io.airbyte.scheduler.worker_run_factories.WorkerRunFactory;
-import io.airbyte.workers.Worker;
+import io.airbyte.scheduler.worker_run.CheckConnectionWorkerRunFactory;
+import io.airbyte.scheduler.worker_run.DiscoverWorkerRunFactory;
+import io.airbyte.scheduler.worker_run.GetSpecWorkerRunFactory;
+import io.airbyte.scheduler.worker_run.SchedulerWorkerRunWithEnvironmentFactory;
+import io.airbyte.scheduler.worker_run.SyncWorkerRunFactories.ResetConnectionWorkerRunFactory;
+import io.airbyte.scheduler.worker_run.SyncWorkerRunFactories.SyncWorkerRunFactory;
+import io.airbyte.scheduler.worker_run.WorkerRunFactory;
 import io.airbyte.workers.process.ProcessBuilderFactory;
-import io.airbyte.workers.wrappers.JobOutputCheckConnectionWorker;
-import io.airbyte.workers.wrappers.JobOutputDiscoverSchemaWorker;
-import io.airbyte.workers.wrappers.JobOutputGetSpecWorker;
-import io.airbyte.workers.wrappers.JobOutputSyncWorker;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
